Bank account upd

+++ NET.S.2018.Brahinets.08/BankAccountLogic/BankAccountsFactory.cs
@@ -9,22 +9,22 @@ namespace BankAccountLogic
 {
     public class BankAccountsFactory:IBankAccountFactory
     {
-        public BankAccount CreateBaseBankAccount(int accountNumber, Client owner, int balance = 0, int bonusPoints = 0)
+        public BankAccount CreateBaseBankAccount(int accountNumber, Client owner, Decimal balance = 0, int bonusPoints = 0)
         {
             return new BaseAccount(accountNumber, owner, balance, bonusPoints);
         }
 
-        public BankAccount CreateSilverBankAccount(int accountNumber, Client owner, int balance = 0, int bonusPoints = 0)
+        public BankAccount CreateSilverBankAccount(int accountNumber, Client owner, Decimal balance = 0, int bonusPoints = 0)
         {
             return new SilverAccount(accountNumber, owner, balance, bonusPoints);
         }
 
-        public BankAccount CreateGoldBankAccount(int accountNumber, Client owner, int balance = 0, int bonusPoints = 0)
+        public BankAccount CreateGoldBankAccount(int accountNumber, Client owner, Decimal balance = 0, int bonusPoints = 0)
         {
             return new GoldAccount(accountNumber, owner, balance, bonusPoints);
         }
 
-        public BankAccount CreatePlatinumBankAccount(int accountNumber, Client owner, int balance = 0, int bonusPoints = 0)
+        public BankAccount CreatePlatinumBankAccount(int accountNumber, Client owner, Decimal balance = 0, int bonusPoints = 0)
         {
             return new PlatinumAccount(accountNumber, owner, balance, bonusPoints);
         }
